requirements.txt
django==3.2.9
djangorestframework==3.13.1
drf-yasg==1.20.0
psycopg2-binary==2.9.3
whitenoise==6.0.0